TFP-3087: Fjerner behovet for AvklarFaktaTillegsopplysningerBekreftelse (#235)
* Fjerner behovet for AvklarFaktaTillegsopplysningerBekreftelse
Fikser deploy av test rapport.
Legger til TODOs for VerdikjedeForeldrepenger.

* Fjerner resten av AvklarFaktaTillegsopplysningerBekreftelse

+++ src/main/java/no/nav/foreldrepenger/autotest/klienter/fpsak/behandlinger/dto/aksjonspunktbekreftelse/VurderBeregnetInntektsAvvikBekreftelse.java
@@ -15,8 +15,8 @@ public class VurderBeregnetInntektsAvvikBekreftelse extends AksjonspunktBekrefte
         super();
     }
 
-    public VurderBeregnetInntektsAvvikBekreftelse leggTilInntekt(Integer inntekt, Long andelsnr) {
-        inntektPrAndelList.add(new InntektPrAndel(inntekt, andelsnr));
+    public VurderBeregnetInntektsAvvikBekreftelse leggTilInntekt(Integer inntekt, Integer andelsnr) {
+        inntektPrAndelList.add(new InntektPrAndel(inntekt, andelsnr.longValue()));
         return this;
     }
